requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
scikit-image>=0.18.0
pillow>=8.0.0
requests>=2.25.0
tqdm>=4.60.0